+++ packages/client-auth/src/lib/config.ts
@@ -59,7 +59,7 @@ export function serverURLFromOptions(options: SDKOptions): URL | null {
 export const SDK_METADATA = {
     language: "typescript",
     openapiDocVersion: "0.0.0",
-    sdkVersion: "1.0.0-alpha.15",
-    genVersion: "2.380.1",
-    userAgent: "speakeasy-sdk/typescript 1.0.0-alpha.15 2.380.1 0.0.0 @formance/sdk-auth",
+    sdkVersion: "1.0.0-alpha.16",
+    genVersion: "2.382.0",
+    userAgent: "speakeasy-sdk/typescript 1.0.0-alpha.16 2.382.0 0.0.0 @formance/sdk-auth",
 } as const;